tests: drop a cast GCC calls useless and clang cannot see
Four legs failed at Build:

  tests/test_strand_race.cpp:374:60: error: useless cast to type 'uint64_t'
      {aka 'long unsigned int'} [-Werror=useless-cast]

`slot` is `std::size_t`, so `100 + slot` is already `uint64_t` on LP64 and the
`static_cast` is an identity conversion. `-Wuseless-cast` is **GCC-only** --
clang has no such warning -- which is why a clang-verified branch passed every
clang leg and failed every GCC one. The mirror image of morph#661, where the
WASM leg's older clang was the only thing that could see a `-Wshadow` the
project suppresses elsewhere.

Dropping the cast rather than silencing the warning: brace-init of `ModelId`
from `std::size_t` is non-narrowing where `size_t` is 64-bit and a widening
conversion where it is 32-bit, so the removal is portable and the intent --
"a distinct non-zero id per slot" -- is unchanged.

Verified both directions with the warning that fired:
  fixed form -> g++ -Werror=useless-cast ... exit 0
  old form   -> the same command reports the same 1 diagnostic
and the full `gcc-debug` build of `morph_tests` now links, 0 errors, where it
previously stopped at this file.

+++ tests/test_strand_race.cpp
@@ -371,7 +371,7 @@ TEST_CASE("StrandExecutor recycles a drained strand under the key that asked for
                 producers.emplace_back([&, slot] {
                     // Distinct, non-zero ids: 0 is `ModelId`'s reserved
                     // "unbound" sentinel.
-                    morph::exec::detail::ModelId const key{static_cast<uint64_t>(100 + slot)};
+                    morph::exec::detail::ModelId const key{100 + slot};
                     for (int round = 0; round < kRounds; ++round) {
                         for (int post = 0; post < kBurst; ++post) {
                             int const seq = (round * kBurst) + post;
